migrated from manual gcs sync thread to /gcs directory in for syncing files
