--- constellaxion/models/tinyllama_1b/model_infra.json
@@ -1,14 +1,0 @@
-{
-  "container_uri": "europe-docker.pkg.dev/vertex-ai/training/tf-gpu.2-14.py310:latest",
-  "requirements": [
-    "trl",
-    "transformers",
-    "dataset",
-    "peft",
-    "google-cloud-storage"
-  ],
-  "machine_type": "g2-standard-4",
-  "accelerator_type": "NVIDIA_L4",
-  "accelerator_count": 1,
-  "replica_count": 1
-}
